Remove unused `requestId` parameter from `setCacheStatus`
Setting the cache status for a page in dev mode is a sequential
operation that only requires the HTML request ID to identify WebSocket
clients across different browser tabs/windows. This is different from
connecting the React debug channel (after which the cache status
handling was modeled), which requires both the HTML request ID as well
as the request ID to correctly associate debug chunks within a client
with the matching request, which might be for the HTML document, a
client-side navigation, or a server function call.