fix isDynamicRSC condition when deployed (#85919)
Next.js relies on on the Resume Data Cache to prevent tearing between
the statically prefetchable parts of the UI by seeding the resume with
the same cache as the revalidation/build. The RDC data is stashed in the
postponed state which is attached to all of the different outputs that
comprise a page (the segments, the HTML, and the dynamic RSC data).

However, when deployed to targets like Vercel, the revalidation was
incorrectly bypassing the cache for the RSC entry, because it thought
that it was a resume. This led to a drift in cached values, because the
revalidation would cause the main RSC entries RDC data to contain a
different value than what was used for the segment prefetches & HTML ISR
entry.

Fixes NAR-493

Co-authored-by: JJ Kasper <jj@jjsweb.site>