Turbopack: remove duplicate traversal implementations (#85853)
We had two implementations of various traversals: first for the SingleModuleGraph, and then also the more general version that can cross module graphs. Let's just use the second for everything (with just a single graph in the case of `SingleModuleGraph`)

This turned out to be a much bigger change than I had anticipated, but this was long overdue anyway.