update routes-manifest to include whether app has pages routes (#86051)
The Vercel CLI produces a lot of `/_next/data` routes regardless of if
pages router is in-use. There's a lot of different metadata we emit but
nothing concretely tells us if an app has pages router enabled. For app
router, we rely on the presence of `app-paths-manifest`, but in pages
router, we always emit a `pages-manifest`.

Rather than trying to cobble together something in the CLI, this updates
the routes manifest to explicitly output the app type (pages, app, or
hybrid).