Turbopack: Refactor output assets to allow to lazy compute output assets (#85753)
### What?

Refactor the OutputAssets of chunk groups to not eagerly return all referenced output assets from async loaders, but instead return a list of references that can be followed to get the full list.
That solves a problem that will happen when the async loaders create cycles in the output graph.