fix: cacheMaxMemorySize should not disable dev HMR cache (#86164)
When configuring `cacheMaxMemorySize` to disable memory caching
(typically used for setups to adjust production caching behavior,
especially in multi-process self-hosted setups), this inadvertently also
disabled dev HMR caches which are intended as a performance
optimization. As a result of this, the dev console would be spammed with
"Single item size exceeds maxsize".

It's still possible to use this configuration to set a larger cache, but
otherwise it will have a minimum set to the previous default.

Fixes NAR-463
x-ref:
https://github.com/vercel/next.js/pull/85153#issuecomment-3429636751