Turbopack: avoid creating another async chunk loader where there is already one available (#85817)
### What?

We can avoid creating another async loader, when there is already an async loader available for that module.
The new async loader would technically be smaller, but we don't have a way to choose the smaller one at runtime anyway.

The async loader would be different for nearly every path in the application and would create a lot of ouptut files.