[turbopack] Model `||`, `&&`, and `??` as control flow operators (#85837)
Handle control flow _operators_ in the turbopack analyzer in order to trigger dead code handling and trim dependencies from dead branches.

This was partially implemented but we were not producing the actual effects for them.

I explored adding support for `?.` as well, but it is a massive complexity issue, so for now that is deferred.