[turbopack] Enable the filesystem cache for dev in canary builds (#85940)
Enable the filesystem cache in canary builds in preparation for enabling it by default in the next release.

It can still be disabled by explicitly setting the option to `false`.

Flipping it for all users is deferred to #85975 so as not to interfere with potential patch releases.

Also fix the filesystem caching test that was accidentally disabled in #84632 due to the file rename.