Match behavior of baseline-browser-mapping with caniuse-lite (#86653)
## What?

Similar to `caniuse-lite` there's a new package in `browserslist` that
also has "data is outdated" message. In order to handle it in the same
way as `caniuse-lite` I've marked it external and added it as a
dependency of Next.js when installing.

I'm still not happy about this. This log is not super helpful and would
preferably be disabled. Alternatively we can switch to loading
browserslist using the rust-side browserslist implementation.

Warning message:

```
[baseline-browser-mapping] The data in this module is over two months old.  To ensure accurate Baseline data, please update: `npm i baseline-browser-mapping@latest -D`
```