Turbopack: chore: Remove re-exports from turbopack crate (#85791)
I don't think these specific re-exports are too problematic, but we've moved away from the pattern of a few top-level crates re-exporting all of the rest of turbopack because in some cases it can make compilation performance much worse in the way that it structures the crate dependency graph.

Re-exports of "internal" crates can be useful in controlling what parts of the API is and isn't "public", but these re-exports aren't really doing that.